refactor: No longer treating Glob as optional.
Since the Glob package is in the "dependencies" of the project package.json,
it's automatically downloaded by NPM when a user installs it. The same
can be said for PNPM and Yarn, because they follow NPM conventions.

It's not that big of a dependency to require anyway.
This should be the case until one of the following happens:
 * Upgrade to ES6 imports, use deasync to import the module
   using the "import()" function, and move the Glob dependency
   into the "optionalDependencies" field.
 * Upgrade to ES6 imports, use my soon to be released special
   imports library, and move the Glob dependency.
 * Or you compile glob into the resulting library using a
   system like Rollup.
 * Or you just move the glob dependency to the "optionalDependencies"
   field.

My point is treating glob as optional is pointless when your package.json
doesn't reflect that. And since it's not invasive to your users at all
to make it fully required, doing so is good for the project because it
enables us to migrate to ES6 import syntax and maintain standard.