chore: Tidyed the repository directory structure.
This is a more extensible format, project compilation can be
served and processed into a new folder apart from the sources,
sources are separate from the configuration files, and
standalone executables are allocated their own location as well.

The cli folder could be renamed, but currently holds our client executables.
The source folder is explicitly for source files..
The root of the repository is for configuration files ONLY, with the exception of the README and LICENSE files.