refactor: Changed the callback style to lambda expression within the repo.
It just looks better and is easier to maintain IMO.

I also went ahead and moved:
 * The imports to three lines below the module header,
 * global variable assignments to two lines beneath the imports,
 * global function definitions to three lines below the globals vars,
 * any execute-on-import (mainlike) code to below the function defs,
 * and any exports to the very bottom.

This style maintains standard import and exports via ES6, and includes
style changes that (could) benefit readablity as a whole. I took
care to add blank lines to the code sparingly, to indicate groups
of action which make a cohesive unit; with the intent that they would
be easier to visually parse.

Overall, the inspiration for my styling comes from:
  https://www.python.org/dev/peps/pep-0008

However I explicitly like to differ when it comes to tabs as
indentation; which you could tell by reading my previous commits.
Spaces enforce a layout without care for control characters and
their potential to put more power into the hands of the reader.
Different people understand code easier at different indentation
levels.

The imports doesn't need to be so far below the module header,
but I've been keeping it at that distance for a while in my own
projects. You can fix that to your preference later.